First page "démarrer" delete and breadcrumb clickable added for cooking challenge
