--- src/pages/lancer-un-defi/culinaire/5.tsx
@@ -1,167 +1,0 @@
-import Image from 'next/image';
-import Link from 'next/link';
-import { useRouter } from 'next/router';
-import React from 'react';
-
-import Backdrop from '@mui/material/Backdrop';
-import Button from '@mui/material/Button';
-import CircularProgress from '@mui/material/CircularProgress';
-import Grid from '@mui/material/Grid';
-
-import { isDefi } from 'src/activity-types/anyActivity';
-import { isCooking, getDefi, DEFI } from 'src/activity-types/defi.constants';
-import type { CookingDefiData } from 'src/activity-types/defi.types';
-import { Base } from 'src/components/Base';
-import { StepsButton } from 'src/components/StepsButtons';
-import { Steps } from 'src/components/Steps';
-import ActivityLink from 'src/components/activities/Link';
-import { Activities } from 'src/components/activities/List';
-import { ContentView } from 'src/components/activities/content/ContentView';
-import { REACTIONS } from 'src/components/activities/utils';
-import { EditButton } from 'src/components/buttons/EditButton';
-import { ActivityContext } from 'src/contexts/activityContext';
-import { useActivity } from 'src/services/useActivity';
-import { ActivityStatus } from 'types/activity.type';
-
-const DefiStep5 = () => {
-  const router = useRouter();
-  const { activity, save } = React.useContext(ActivityContext);
-  const { activity: responseActivity } = useActivity(activity?.responseActivityId ?? -1);
-  const [isLoading, setIsLoading] = React.useState(false);
-
-  const data = (activity?.data as CookingDefiData) || null;
-  const isEdit = activity !== null && activity.id !== 0 && activity.status !== ActivityStatus.DRAFT;
-
-  React.useEffect(() => {
-    if (activity === null && !('activity-id' in router.query) && !sessionStorage.getItem('activity')) {
-      router.push('/lancer-un-defi');
-    } else if (activity && (!isDefi(activity) || (isDefi(activity) && !isCooking(activity)))) {
-      router.push('/lancer-un-defi');
-    }
-  }, [activity, router]);
-
-  const onPublish = async () => {
-    setIsLoading(true);
-    const success = await save(true);
-    if (success) {
-      router.push('/lancer-un-defi/success');
-    }
-    setIsLoading(false);
-  };
-
-  if (data === null || activity === null || !isDefi(activity) || (isDefi(activity) && !isCooking(activity))) {
-    return <div></div>;
-  }
-
-  return (
-    <Base>
-      <div style={{ width: '100%', padding: '0.5rem 1rem 1rem 1rem' }}>
-        <Steps steps={(isEdit ? [] : ['Démarrer']).concat(['Votre plat', 'La recette', 'Le défi', 'Prévisualisation'])} activeStep={isEdit ? 3 : 4} />
-        <div className="width-900">
-          <h1>Pré-visualisez votre défi{!isEdit && ', et publiez-la'}</h1>
-          <p className="text" style={{ fontSize: '1.1rem' }}>
-            Voici la pré-visualisation de votre défi.
-            {isEdit
-              ? " Vous pouvez le modifier à l'étape précédente, et enregistrer vos changements ici."
-              : ' Vous pouvez le modifier, et quand vous êtes prêts : publiez-le dans votre village-monde !'}
-          </p>
-          {isEdit ? (
-            <div style={{ width: '100%', display: 'flex', justifyContent: 'space-between', margin: '1rem 0' }}>
-              <Link href="/lancer-un-defi/culinaire/4" passHref>
-                <Button component="a" color="secondary" variant="contained" href="/lancer-un-defi/culinaire/4">
-                  {"Modifier à l'étape précédente"}
-                </Button>
-              </Link>
-              <Button variant="outlined" color="primary" onClick={onPublish}>
-                Enregistrer les changements
-              </Button>
-            </div>
-          ) : (
-            <div style={{ width: '100%', textAlign: 'right', margin: '1rem 0' }}>
-              <Button variant="outlined" color="primary" onClick={onPublish}>
-                Publier
-              </Button>
-            </div>
-          )}
-
-          {!isEdit && activity.responseActivityId === null && <ActivityLink url={`/lancer-un-defi/culinaire/1?edit=${activity.id}`} />}
-          {responseActivity !== null && (
-            <>
-              <span className="text text--small text--success">Défi en réaction à {REACTIONS[responseActivity.type]}</span>
-              <div className="preview-block">
-                {!isEdit && (
-                  <EditButton
-                    onClick={() => {
-                      router.push(`/lancer-un-defi/culinaire/1?edit=${activity.id}`);
-                    }}
-                    status={'success'}
-                    style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
-                  />
-                )}
-                <Activities activities={[responseActivity]} noButtons />
-              </div>
-            </>
-          )}
-
-          <span className="text text--small text--success">Description de votre plat</span>
-          <div className="preview-block">
-            <EditButton
-              onClick={() => {
-                router.push('/lancer-un-defi/culinaire/2');
-              }}
-              status={'success'}
-              style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
-            />
-            <div className="text-center" style={{ marginBottom: '1rem' }}>
-              <h3>{data.name}</h3>
-            </div>
-            <Grid container spacing={2}>
-              {data.image && (
-                <Grid item xs={12} md={4}>
-                  <div style={{ width: '100%', height: '100%', minHeight: '200px', position: 'relative' }}>
-                    <Image layout="fill" objectFit="contain" alt="image du plat" unoptimized src={data.image} />
-                  </div>
-                </Grid>
-              )}
-              <Grid item xs={12} md={data.image ? 8 : 12}>
-                <p>{data.history}</p>
-                <p>{data.explanation}</p>
-              </Grid>
-            </Grid>
-          </div>
-
-          <span className="text text--small text--success">La recette</span>
-          <div className="preview-block">
-            <EditButton
-              onClick={() => {
-                router.push('/lancer-un-defi/culinaire/3');
-              }}
-              status={'success'}
-              style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
-            />
-            <ContentView content={activity.content} />
-          </div>
-
-          <span className="text text--small text--success">Le défi lancé aux Pélicopains</span>
-          <div className="preview-block">
-            <EditButton
-              onClick={() => {
-                router.push('/lancer-un-defi/culinaire/4');
-              }}
-              status={'success'}
-              style={{ position: 'absolute', top: '0.5rem', right: '0.5rem' }}
-            />
-            Votre défi : {getDefi(DEFI.COOKING, data)}
-          </div>
-
-          <StepsButton prev="/lancer-un-defi/culinaire/4" />
-        </div>
-      </div>
-      <Backdrop style={{ zIndex: 2000, color: 'white' }} open={isLoading}>
-        <CircularProgress color="inherit" />
-      </Backdrop>
-    </Base>
-  );
-};
-
-export default DefiStep5;
